Changed package name of test classes and jars
