+++ src/test/resources-src/multi-jdk/src/main/java/org/vafer/multijdk/AbstractJavaVersion.java
@@ -1,4 +1,4 @@
-package nl.basjes.maven.multijdk;
+package org.vafer.multijdk;
 /*
  * Copyright 2010-2023 The jdependency developers.
  *
@@ -14,8 +14,19 @@ package nl.basjes.maven.multijdk;
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-public class Unused {
-    public String doSomething() {
-        return "java 17";
+
+public abstract class AbstractJavaVersion {
+
+    public abstract String getCodeVersion();
+    public abstract String getJavaVersion();
+
+    public String getJavaMajorVersion() {
+        String version = getJavaVersion();
+        String[] versionElements = version.split("\\.");
+        String majorVersion = versionElements[0];
+        if ("1".equals(majorVersion)) {
+            majorVersion = versionElements[1];
+        }
+        return majorVersion;
     }
 }
