+++ src/test/resources-src/multi-jdk/src/main/java/org/vafer/multijdk/JavaVersion.java
@@ -1,4 +1,4 @@
-package nl.basjes.maven.multijdk;
+package org.vafer.multijdk;
 /*
  * Copyright 2010-2023 The jdependency developers.
  *
@@ -15,8 +15,12 @@ package nl.basjes.maven.multijdk;
  * limitations under the License.
  */
 
-public class SpecificToJava11 {
-    public String doSomething() {
-        return "java";
+public class JavaVersion extends AbstractJavaVersion {
+    public String getCodeVersion() {
+        return "8";
+    }
+
+    public String getJavaVersion() {
+        return System.getProperty("java.version");
     }
 }
