+++ src/test/resources-src/multi-jdk/src/main/java/org/vafer/multijdk/Main.java
@@ -1,4 +1,4 @@
-package nl.basjes.maven.multijdk;
+package org.vafer.multijdk;
 /*
  * Copyright 2010-2023 The jdependency developers.
  *
@@ -15,8 +15,12 @@ package nl.basjes.maven.multijdk;
  * limitations under the License.
  */
 
-public class App {
-    public String doSomething() {
-        return new SpecificToJava17().doSomething() + " " + new OnlyUsedInJava17().doSomething();
+public class Main {
+    public static void main(String[] args) {
+        JavaVersion javaVersion = new JavaVersion();
+        System.out.println("Java detect: " + javaVersion.getCodeVersion());
+        System.out.println("Java major : " + javaVersion.getJavaMajorVersion());
+        System.out.println("App code   : " + new App().doSomething());
+        System.out.println("Unused     : " + new Unused().doSomething());
     }
 }
